- rename all mySerial to grblSerial
- add and use phi_interfaces for rotary encoder
- change intervall to 500ms and 1000ms for threads
- send reset if connected to grbl for UniversalGcodeSender
- add call_button

+++ serialspy/Visual Micro/.serialspy.vsarduino.h
@@ -16,7 +16,7 @@ void setinterval(String line);
 void resetDevice();
 void memoryfree();
 void about();
-void call_button(int number);
+void call_button(byte number);
 void call_button_1();
 void call_button_2();
 void call_button_3();
